C-WCOW: Add SecurityPolicy tests
